+++ volmdlr/surfaces.py
@@ -192,10 +192,7 @@ class Surface2D(PhysicalObject):
                'segments': npy.array(segments).reshape((-1, 2)),
                }
         triagulation = triangle_lib.triangulate(tri, tri_opt)
-        triangles = triagulation['triangles'].tolist()
-        number_points = triagulation['vertices'].shape[0]
-        points = [display.Node2D(*triagulation['vertices'][i, :]) for i in range(number_points)]
-        return display.DisplayMesh2D(points, triangles=triangles)
+        return display.Mesh2D(vertices=triagulation['vertices'], triangles=triagulation['triangles'])
 
     def triangulation(self, number_points_x: int = 15, number_points_y: int = 15):
         """
@@ -206,12 +203,12 @@ class Surface2D(PhysicalObject):
         :param number_points_y: Number of discretization points in y direction.
         :type number_points_y: int
         :return: The triangulated surface as a display mesh.
-        :rtype: :class:`volmdlr.display.DisplayMesh2D`
+        :rtype: :class:`volmdlr.display.Mesh2D`
         """
         area = self.bounding_rectangle().area()
         tri_opt = "p"
-        if math.isclose(area, 0., abs_tol=1e-8):
-            return display.DisplayMesh2D([], triangles=[])
+        if math.isclose(area, 0., abs_tol=1e-12):
+            return None
 
         triangulates_with_grid = number_points_x > 0 and number_points_y > 0
         discretize_line = number_points_x > 0 or number_points_y > 0
@@ -288,10 +285,7 @@ class Surface2D(PhysicalObject):
                'holes': npy.array(holes).reshape((-1, 2))
                }
         triangulation = triangle_lib.triangulate(tri, tri_opt)
-        triangles = triangulation['triangles'].tolist()
-        number_points = triangulation['vertices'].shape[0]
-        points = [volmdlr.Point2D(*triangulation['vertices'][i, :]) for i in range(number_points)]
-        return display.DisplayMesh2D(points, triangles=triangles)
+        return display.Mesh2D(vertices=triangulation['vertices'], triangles=triangulation['triangles'])
 
     def split_by_lines(self, lines):
         """
@@ -772,45 +766,60 @@ class Surface3D(DessiaObject):
                 primitives2d[0] = self.fix_undefined_brep_with_neighbors(primitives2d[0], primitives2d[-1],
                                                                          primitives2d[1])
                 primitives_mapping[primitives2d[0]] = primitives_mapping.pop(old_primitive)
+        self._helper_repair_primitives_periodicity(primitives2d, primitives_mapping,
+                                                   [x_periodicity, y_periodicity], tol)
+        if self.__class__.__name__ in ("SphericalSurface3D", "ConicalSurface3D", "RevolutionSurface3D"):
+            delta = primitives2d[-1].end - primitives2d[0].start
+            if (math.isclose(abs(delta.x), x_periodicity, abs_tol=tol) or
+                    math.isclose(abs(delta.y), y_periodicity, abs_tol=tol)):
+                last_end_3d = self.point2d_to_3d(primitives2d[-1].end)
+                first_start_3d = self.point2d_to_3d(primitives2d[0].start)
+                if last_end_3d.is_close(first_start_3d, 1e-6) and not self.is_singularity_point(last_end_3d):
+                    old_primitive = primitives2d[0]
+                    primitives2d[0] = primitives2d[0].translation(delta)
+                    primitives_mapping[primitives2d[0]] = primitives_mapping.pop(old_primitive)
+                    self._helper_repair_primitives_periodicity(primitives2d, primitives_mapping,
+                                                               [x_periodicity, y_periodicity], tol)
+        self.check_parametric_contour_end(primitives2d, tol)
+
+    def _helper_repair_primitives_periodicity(self, primitives2d, primitives_mapping, periodicities, tol):
+        """Helper function to repair_primitives_periodicity."""
+        x_periodicity, y_periodicity = periodicities
         i = 1
         while i < len(primitives2d):
-            previous_primitive = primitives2d[i - 1]
-            current_primitive = primitives2d[i]
-            delta = previous_primitive.end - current_primitive.start
+            delta = primitives2d[i - 1].end - primitives2d[i].start
             distance = delta.norm()
-            is_connected = math.isclose(distance, 0, abs_tol=tol)
 
-            if not is_connected:
-                if math.isclose(current_primitive.length(), x_periodicity, abs_tol=tol) or \
-                        math.isclose(current_primitive.length(), y_periodicity, abs_tol=tol):
-                    delta_end = previous_primitive.end - current_primitive.end
+            if not math.isclose(distance, 0, abs_tol=tol):
+                if math.isclose(primitives2d[i].length(), x_periodicity, abs_tol=tol) or \
+                        math.isclose(primitives2d[i].length(), y_periodicity, abs_tol=tol):
+                    delta_end = primitives2d[i - 1].end - primitives2d[i].end
                     delta_min_index, _ = min(enumerate([distance, delta_end.norm()]), key=lambda x: x[1])
                     if self.is_undefined_brep(primitives2d[i]):
-                        repair_undefined_brep(self, primitives2d, primitives_mapping, i, previous_primitive)
-                    elif self.is_singularity_point(self.point2d_to_3d(previous_primitive.end)) and \
-                            self.is_singularity_point(self.point2d_to_3d(current_primitive.start)):
-                        self.repair_singularity(primitives2d, i, previous_primitive)
-                    elif current_primitive.end.is_close(previous_primitive.end, tol=tol):
+                        repair_undefined_brep(self, primitives2d, primitives_mapping, i, primitives2d[i - 1])
+                    elif self.is_singularity_point(self.point2d_to_3d(primitives2d[i - 1].end)) and \
+                            self.is_singularity_point(self.point2d_to_3d(primitives2d[i].start)):
+                        self.repair_singularity(primitives2d, i, primitives2d[i - 1])
+                    elif primitives2d[i].end.is_close(primitives2d[i - 1].end, tol=tol):
                         self.repair_reverse(primitives2d, primitives_mapping, i)
                     elif delta_min_index == 0:
                         self.repair_translation(primitives2d, primitives_mapping, i, delta)
                     else:
                         old_primitive = primitives2d[i]
-                        new_primitive = current_primitive.reverse()
+                        new_primitive = primitives2d[i].reverse()
                         primitives2d[i] = new_primitive.translation(delta_end)
                         primitives_mapping[primitives2d[i]] = primitives_mapping.pop(old_primitive)
 
-                elif current_primitive.end.is_close(previous_primitive.end, tol=tol):
+                elif primitives2d[i].end.is_close(primitives2d[i - 1].end, tol=tol):
                     self.repair_reverse(primitives2d, primitives_mapping, i)
                 elif self.is_undefined_brep(primitives2d[i]):
-                    repair_undefined_brep(self, primitives2d, primitives_mapping, i, previous_primitive)
-                elif self.is_singularity_point(self.point2d_to_3d(previous_primitive.end), tol=1e-5) and \
-                        self.is_singularity_point(self.point2d_to_3d(current_primitive.start), tol=1e-5):
-                    self.repair_singularity(primitives2d, i, previous_primitive)
+                    repair_undefined_brep(self, primitives2d, primitives_mapping, i, primitives2d[i - 1])
+                elif self.is_singularity_point(self.point2d_to_3d(primitives2d[i - 1].end), tol=1e-5) and \
+                        self.is_singularity_point(self.point2d_to_3d(primitives2d[i].start), tol=1e-5):
+                    self.repair_singularity(primitives2d, i, primitives2d[i - 1])
                 else:
                     self.repair_translation(primitives2d, primitives_mapping, i, delta)
             i += 1
-        self.check_parametric_contour_end(primitives2d, tol)
 
     def check_parametric_contour_end(self, primitives2d, tol):
         """Helper function to repair_primitives_periodicity."""
@@ -821,7 +830,7 @@ class Surface3D(DessiaObject):
         if not is_connected and self.is_singularity_point(self.point2d_to_3d(previous_primitive.end)) and \
                 self.is_singularity_point(self.point2d_to_3d(primitives2d[0].start)):
             primitives2d.append(edges.LineSegment2D(previous_primitive.end, primitives2d[0].start,
-                                                name="construction"))
+                                                    name="construction"))
 
     @staticmethod
     def repair_singularity(primitives2d, i, previous_primitive):
@@ -1616,6 +1625,31 @@ class Plane3D(Surface3D):
         """
         return point2d.to_3d(self.frame.origin, self.frame.u, self.frame.v)
 
+    def parametric_points_to_3d(self, points: NDArray[npy.float64]) -> NDArray[npy.float64]:
+        """
+        Transform parametric coordinates to 3D points on the plane.
+
+        Given a set of parametric coordinates `(u, v)` representing points on the surface,
+        this method returns the corresponding 3D points on the plane.
+
+        :param points: Parametric coordinates in the form of a numpy array with shape (n, 2),
+                       where `n` is the number of points, and each row corresponds to `(u, v)`.
+        :type points: numpy.ndarray[npy.float64]
+
+        :return: Array of 3D points representing the plane in Cartesian coordinates.
+        :rtype: numpy.ndarray[npy.float64]
+        """
+        center = npy.array(self.frame.origin)
+        x = npy.array([self.frame.u[0], self.frame.u[1], self.frame.u[2]])
+        y = npy.array([self.frame.v[0], self.frame.v[1], self.frame.v[2]])
+
+        points = points.reshape(-1, 2, 1)
+
+        u_values = points[:, 0]
+        v_values = points[:, 1]
+
+        return center + u_values * x + v_values * y
+
     def point3d_to_2d(self, point3d):
         """
         Converts a 3D point into a 2D parametric point.
@@ -2296,13 +2330,19 @@ class CylindricalSurface3D(PeriodicalSurface):
             return True
         return False
 
-    def get_generatrices(self, length: float = 1, number_lines: int = 30):
+    def get_generatrices(self, number_lines: int = 30, length: float = 1):
         """
-        Gets the Cylindrical surface's Line generatrices.
+        Retrieve line segments representing the generatrices of a cylinder.
 
-        :param number_lines: number of lines
-        :param length: the length used to determine the lines' length.
-        :return: list of cylindrical surface's circular generatrices.
+        Generates a specified number of line segments along the surface of the cylinder,
+        each representing a generatrix.
+
+        :param number_lines: The number of generatrices to generate. Default is 30
+        :type number_lines: int
+        :param length: The length of the cylinder along the z-direction. Default is 1.
+        :type length: float
+        :return: A list of LineSegment3D instances representing the generatrices of the cylinder.
+        :rtype: List[LineSegment3D]
         """
         list_generatrices = []
         for i in range(number_lines):
@@ -2315,11 +2355,17 @@ class CylindricalSurface3D(PeriodicalSurface):
 
     def get_circle_generatrices(self, number_circles: int = 10, length: float = 1.0):
         """
-        Gets the Cylindrical surface's circular generatrices.
+        Retrieve circles representing the generatrices of a cylinder.
 
-        :param number_circles: number of circles
-        :param length: the length used to determine the circles creation range.
-        :return: list of cylindrical surface's circular generatrices.
+        Generates a specified number of circles along the surface of the cylinder,
+        each representing a generatrix.
+
+        :param number_circles: The number of generatrices to generate. Default is 10
+        :type number_circles: int
+        :param length: The length of the cylinder along the z-direction. Default is 1.
+        :type length: float
+        :return: A list of Circle3D instances representing the generatrices of the cylinder.
+        :rtype: List[Circle3D]
         """
         circles = []
         for j in range(number_circles):
@@ -2353,7 +2399,7 @@ class CylindricalSurface3D(PeriodicalSurface):
             length = self.radius
 
         self.frame.plot(ax=ax, color=edge_style.color, ratio=self.radius)
-        for edge in self.get_generatrices(length, nlines):
+        for edge in self.get_generatrices(nlines, length):
             edge.plot(ax=ax, edge_style=edge_style)
 
         circles = self.get_circle_generatrices(ncircles, length)
@@ -2609,7 +2655,7 @@ class CylindricalSurface3D(PeriodicalSurface):
         :return: list of intersecting curves.
         """
         def _list_generatrices_intersections(surface, other_surface):
-            linesegments = other_surface.get_generatrices(2, 50)
+            linesegments = other_surface.get_generatrices(50, 2)
             all_generatrices_intersecting = True
             lists_intersections = [[], []]
             for generatrix in linesegments:
@@ -2779,7 +2825,7 @@ class CylindricalSurface3D(PeriodicalSurface):
         :param cylindricalsurface: other Cylindrical surface 3d.
         :return: points of intersections.
         """
-        cyl_generatrices = self.get_generatrices(self.radius*10, 200) +\
+        cyl_generatrices = self.get_generatrices(200, self.radius*10) +\
                            self.get_circle_generatrices(200, self.radius*10)
         intersection_points = []
         for gene in cyl_generatrices:
@@ -3255,7 +3301,7 @@ class ToroidalSurface3D(PeriodicalSurface):
         """
         Triangulation.
 
-        :rtype: display.DisplayMesh3D
+        :rtype: display.Mesh3D
         """
         face = self.rectangular_cut(0, volmdlr.TWO_PI, 0, volmdlr.TWO_PI)
         return face.triangulation()
@@ -3617,7 +3663,7 @@ class ToroidalSurface3D(PeriodicalSurface):
             for intersection in intersections:
                 if not intersection.in_list(points_intersections):
                     points_intersections.append(intersection)
-        for edge in cylindrical_surface.get_generatrices(self.outer_radius * 3, 300):
+        for edge in cylindrical_surface.get_generatrices(300, self.outer_radius * 3):
             # \
             #    cylindrical_surface.get_circle_generatrices(72, self.outer_radius * 3):
             intersections = self.line_intersections(edge.line)
@@ -3684,8 +3730,8 @@ class ToroidalSurface3D(PeriodicalSurface):
             points_intersections.extend(intersections)
         point1 = conical_surface.frame.global_to_local_coordinates(volmdlr.Point3D(0, 0, self.bounding_box.zmin))
         point2 = conical_surface.frame.global_to_local_coordinates(volmdlr.Point3D(0, 0, self.bounding_box.zmax))
-        for edge in conical_surface.get_generatrices(self.outer_radius * 3, 300) + \
-                conical_surface.get_circle_generatrices(point1.z, point2.z, 100):
+        for edge in conical_surface.get_generatrices(300, self.outer_radius * 3) + \
+                conical_surface.get_circle_generatrices(100, point1.z, point2.z):
             intersections = self.edge_intersections(edge)
             for point in intersections:
                 if not point.in_list(points_intersections):
@@ -3927,7 +3973,7 @@ class ConicalSurface3D(PeriodicalSurface):
         """Returns u and v bounds."""
         return -math.pi, math.pi, -math.inf, math.inf
 
-    def get_generatrices(self, z: float = 1, number_lines: int = 36):
+    def get_generatrices(self, number_lines: int = 36, z: float = 1):
         """
         Gets Conical Surface 3D generatrix lines.
 
@@ -3953,7 +3999,7 @@ class ConicalSurface3D(PeriodicalSurface):
         circle = curves.Circle3D(i_frame, radius)
         return circle
 
-    def get_circle_generatrices(self, z1, z2, number_circles: int):
+    def get_circle_generatrices(self, number_circles: int, z1, z2):
         """
         Get circles generatrix of the cone.
 
@@ -3978,8 +4024,9 @@ class ConicalSurface3D(PeriodicalSurface):
             fig = plt.figure()
             ax = fig.add_subplot(111, projection='3d')
 
-        line_generatrices = self.get_generatrices(z, 36)
-        circle_generatrices = self.get_circle_generatrices(0, z, 50)
+        line_generatrices = self.get_generatrices(36, z)
+        circle_generatrices = self.get_circle_generatrices(50, 0, z)
+
         for edge in line_generatrices + circle_generatrices:
             edge.plot(ax, edge_style)
         return ax
@@ -4078,6 +4125,35 @@ class ConicalSurface3D(PeriodicalSurface):
         if abs(theta) < 1e-9:
             theta = 0.0
         return volmdlr.Point2D(theta, z)
+
+    def parametric_points_to_3d(self, points: NDArray[npy.float64]) -> NDArray[npy.float64]:
+        """
+        Transform parametric coordinates to 3D points on the conical surface.
+
+        Given a set of parametric coordinates `(u, v)` representing points on the surface,
+        this method returns the corresponding 3D points on the conical surface.
+
+        :param points: Parametric coordinates in the form of a numpy array with shape (n, 2),
+                       where `n` is the number of points, and each row corresponds to `(u, v)`.
+        :type points: numpy.ndarray[npy.float64]
+
+        :return: Array of 3D points representing the conical surface in Cartesian coordinates.
+        :rtype: numpy.ndarray[npy.float64]
+        """
+        center = npy.array(self.frame.origin)
+        x = npy.array([self.frame.u[0], self.frame.u[1], self.frame.u[2]])
+        y = npy.array([self.frame.v[0], self.frame.v[1], self.frame.v[2]])
+        z = npy.array([self.frame.w[0], self.frame.w[1], self.frame.w[2]])
+
+        points = points.reshape(-1, 2, 1)
+
+        u_values = points[:, 0]
+        v_values = points[:, 1]
+
+        x_component = npy.cos(u_values) * x
+        y_component = npy.sin(u_values) * y
+
+        return center + v_values * math.tan(self.semi_angle) * (x_component + y_component) + v_values * z
 
     def rectangular_cut(self, theta1: float, theta2: float,
                         param_z1: float, param_z2: float, name: str = ''):
@@ -4467,8 +4543,8 @@ class ConicalSurface3D(PeriodicalSurface):
         """
         point1 = self.frame.global_to_local_coordinates(volmdlr.Point3D(0, 0, spherical_surface.bounding_box.zmin))
         point2 = self.frame.global_to_local_coordinates(volmdlr.Point3D(0, 0, spherical_surface.bounding_box.zmax))
-        cone_generatrices = self.get_generatrices(spherical_surface.radius*4, 200) +\
-                            self.get_circle_generatrices(point1.z, point2.z, 200)
+        cone_generatrices = self.get_generatrices(200, spherical_surface.radius*4) +\
+                            self.get_circle_generatrices(200, point1.z, point2.z)
         intersection_points = []
         for gene in cone_generatrices:
             intersections = spherical_surface.edge_intersections(gene)
@@ -4515,8 +4591,8 @@ class ConicalSurface3D(PeriodicalSurface):
         :param conical_surface: other Spherical Surface 3d.
         :return: points of intersections.
         """
-        cone_generatrices = self.get_generatrices(length, max(100, int((length / 2) * 10))) + \
-                            self.get_circle_generatrices(0, length, max(200, int((length / 2) * 20)))
+        cone_generatrices = self.get_generatrices(max(100, int((length / 2) * 10)), length) + \
+                            self.get_circle_generatrices(max(200, int((length / 2) * 20)), 0, length)
         intersection_points = []
         for gene in cone_generatrices:
             intersections = conical_surface.edge_intersections(gene)
@@ -4811,6 +4887,37 @@ class SphericalSurface3D(PeriodicalSurface):
             phi = 0
 
         return volmdlr.Point2D(theta, phi)
+
+    def parametric_points_to_3d(self, points: NDArray[npy.float64]) -> NDArray[npy.float64]:
+        """
+        Transform parametric coordinates to 3D points on the spherical surface.
+
+        Given a set of parametric coordinates `(u, v)` representing points on the surface,
+        this method returns the corresponding 3D points on the spherical surface.
+
+        :param points: Parametric coordinates in the form of a numpy array with shape (n, 2),
+                       where `n` is the number of points, and each row corresponds to `(u, v)`.
+        :type points: numpy.ndarray[npy.float64]
+
+        :return: Array of 3D points representing the spherical surface in Cartesian coordinates.
+        :rtype: numpy.ndarray[npy.float64]
+        """
+        center = npy.array(self.frame.origin)
+        x = npy.array([self.frame.u[0], self.frame.u[1], self.frame.u[2]])
+        y = npy.array([self.frame.v[0], self.frame.v[1], self.frame.v[2]])
+        z = npy.array([self.frame.w[0], self.frame.w[1], self.frame.w[2]])
+
+        points = points.reshape(-1, 2, 1)
+
+        u_values = points[:, 0]
+        v_values = points[:, 1]
+
+        common_term = self.radius * npy.cos(v_values)
+        x_component = npy.cos(u_values) * x
+        y_component = npy.sin(u_values) * y
+        z_component = self.radius * npy.sin(v_values) * z
+
+        return center + common_term * (x_component + y_component) + z_component
 
     def linesegment2d_to_3d(self, linesegment2d):
         """
@@ -5769,6 +5876,34 @@ class ExtrusionSurface3D(Surface3D):
 
         return volmdlr.Point2D(u, v)
 
+    def parametric_points_to_3d(self, points: NDArray[npy.float64]) -> NDArray[npy.float64]:
+        """
+        Transform parametric coordinates to 3D points on the extrusion surface.
+
+        Given a set of parametric coordinates `(u, v)` representing points on the surface,
+        this method returns the corresponding 3D points on the extrusion surface.
+
+        :param points: Parametric coordinates in the form of a numpy array with shape (n, 2),
+                       where `n` is the number of points, and each row corresponds to `(u, v)`.
+        :type points: numpy.ndarray[npy.float64]
+
+        :return: Array of 3D points representing the extrusion surface in Cartesian coordinates.
+        :rtype: numpy.ndarray[npy.float64]
+        """
+        z = npy.array([self.direction[0], self.direction[1], self.direction[2]])
+
+        points = points.reshape(-1, 2, 1)
+
+        u_values = points[:, 0]
+        if self.x_periodicity:
+            u_values[u_values > self.x_periodicity] -= self.x_periodicity
+            u_values[u_values < 0] += self.x_periodicity
+        v_values = points[:, 1]
+
+        points_at_curve = npy.array([self.edge.point_at_abscissa(u) for u in u_values])
+
+        return points_at_curve + v_values * z
+
     def rectangular_cut(self, x1: float = 0.0, x2: float = 1.0,
                         y1: float = 0.0, y2: float = 1.0, name: str = ''):
         """Deprecated method, Use ExtrusionFace3D from_surface_rectangular_cut method."""
@@ -6143,6 +6278,40 @@ class RevolutionSurface3D(PeriodicalSurface):
         point_at_curve = point3d.rotation(self.axis_point, self.axis, -u)
         v = self.edge.abscissa(point_at_curve)
         return volmdlr.Point2D(u, v)
+
+    def parametric_points_to_3d(self, points: NDArray[npy.float64]) -> NDArray[npy.float64]:
+        """
+        Transform parametric coordinates to 3D points on the revolution surface.
+
+        Given a set of parametric coordinates `(u, v)` representing points on the surface,
+        this method returns the corresponding 3D points on the revolution surface.
+
+        :param points: Parametric coordinates in the form of a numpy array with shape (n, 2),
+                       where `n` is the number of points, and each row corresponds to `(u, v)`.
+        :type points: numpy.ndarray[npy.float64]
+
+        :return: Array of 3D points representing the revolution surface in Cartesian coordinates.
+        :rtype: numpy.ndarray[npy.float64]
+        """
+        center = npy.array(self.axis_point)
+        z = npy.array([self.axis[0], self.axis[1], self.axis[2]])
+
+        points = points.reshape(-1, 2, 1)
+
+        u_values = points[:, 0]
+        v_values = points[:, 1]
+        if self.y_periodicity:
+            v_values[v_values > self.y_periodicity] -= self.y_periodicity
+            v_values[v_values < 0] += self.y_periodicity
+
+        cos_u = npy.cos(u_values)
+
+        points_at_curve = npy.array([self.edge.point_at_abscissa(v) for v in v_values])
+        points_at_curve_minus_center = points_at_curve - center
+
+        return (center + points_at_curve_minus_center * cos_u +
+                npy.dot(points_at_curve_minus_center, z).reshape(-1, 1) * z * (1 - cos_u) +
+                npy.cross(z, points_at_curve_minus_center * npy.sin(u_values)))
 
     def rectangular_cut(self, x1: float, x2: float,
                         y1: float, y2: float, name: str = ''):
@@ -7296,7 +7465,7 @@ class BSplineSurface3D(Surface3D):
 
     @staticmethod
     def _find_index_min(matrix_points, point):
-        # Calculate distances
+        """Helper function to find point of minimal distance."""
         distances = npy.linalg.norm(matrix_points - point, axis=1)
 
         return npy.argmin(distances), distances.min()
@@ -7567,6 +7736,23 @@ class BSplineSurface3D(Surface3D):
         x[0] = u
         x[1] = v
         return x
+
+    def parametric_points_to_3d(self, points: NDArray[npy.float64]) -> NDArray[npy.float64]:
+        """
+        Transform parametric coordinates to 3D points on the BSpline surface.
+
+        Given a set of parametric coordinates `(u, v)` representing points on the surface,
+        this method returns the corresponding 3D points on the BSpline surface.
+
+        :param points: Parametric coordinates in the form of a numpy array with shape (n, 2),
+                       where `n` is the number of points, and each row corresponds to `(u, v)`.
+        :type points: numpy.ndarray[npy.float64]
+
+        :return: Array of 3D points representing the BSpline surface in Cartesian coordinates.
+        :rtype: numpy.ndarray[npy.float64]
+        """
+        return npy.array([evaluate_surface(self.data, start=(u, v), stop=(u, v))[0] for u, v in points],
+                         dtype=npy.float64)
 
     def linesegment2d_to_3d(self, linesegment2d):
         """Evaluates the Euclidean form for the parametric line segment."""